댓글 삭제 비밀번호
댓글 삭제할때 비밀번호 받아서 맞춰야 삭제되도록 수정~!

+++ venvproject/h0zzae/home/views.py
@@ -75,9 +75,12 @@ def delete(request,post_id):
 
 def replyDelete(request,reply_id,post_id):
     reply = Reply.objects.get(id=reply_id)
-    reply.delete()
-
-    return redirect('/post/'+str(post_id))
+    print(request.POST.get('confirmPassword', False))
+    passwordConfirm=request.POST.get('confirmPassword', False)
+    if reply.password==passwordConfirm:
+        reply.delete()
+        return redirect('/post/'+str(post_id))
+    return render(request, 'post/delete.html')
 
 def editPost(request, post_id):
     post = Post.objects.get(id=post_id)
